update the carousel and now it can work
