+++ packages/client/src/components/ViewPage/Carousel/Carousel.css
@@ -1,0 +1,11 @@
+.view-carousel {
+  height: 400px;
+  width: 620px;
+}
+img {
+  background-color: #a1b6ec;
+  width: 200px;
+  height: 100px;
+  filter: drop-shadow(20px 20px 40px rgba(0, 0, 0, 0.25));
+  border-radius: 112.5px;
+}
