+++ packages/client/src/components/ViewPage/Carousel/mock.json
@@ -1,0 +1,38 @@
+[
+  {
+    "id": 2,
+    "title": "The Shawshank Redemption",
+    "description": "Over the course of several years, two convicts form a friendship, seeking consolation and, eventually, redemption through basic compassion.",
+    "movie_year": 1994,
+    "image_location": "https://www.themoviedb.org/t/p/w600_and_h900_bestv2/q6y0Go1tsGEsmtFryDOJo3dEmqu.jpg",
+    "price": "80.00",
+    "created_at": "2023-05-10T12:42:17.249Z"
+  },
+  {
+    "id": 3,
+    "title": "A Beautiful Mind",
+    "description": "After John Nash, a brilliant but asocial mathematician, accepts secret work in cryptography, his life takes a turn for the nightmarish.",
+    "movie_year": 2001,
+    "image_location": "https://www.themoviedb.org/t/p/w1280/zwzWCmH72OSC9NA0ipoqw5Zjya8.jpg",
+    "price": "90.00",
+    "created_at": "2023-05-10T12:42:17.249Z"
+  },
+  {
+    "id": 6,
+    "title": "I Can Only Imagine",
+    "description": "The inspiring and unknown true story behind MercyMes beloved, chart-topping song that brings ultimate hope to so many is a gripping reminder of the power of true forgiveness.",
+    "movie_year": 2018,
+    "image_location": "https://www.themoviedb.org/t/p/w1280/veZszwMZu8d3WMU6TJX9sV5w1Y4.jpg",
+    "price": "80.00",
+    "created_at": "2023-05-10T12:42:17.249Z"
+  },
+  {
+    "id": 12,
+    "title": "The Dark Knight",
+    "description": "Batman raises the stakes in his war on crime. With the help of Lt. Jim Gordon and District Attorney Harvey Dent, Batman sets out to dismantle the remaining criminal organizations that plague the streets. The partnership proves to be effective.",
+    "movie_year": 2008,
+    "image_location": "https://www.themoviedb.org/t/p/w1280/z4XwvnDZSm40PCdg9vFJmo0W9ri.jpg",
+    "price": "120.00",
+    "created_at": "2023-05-10T12:42:17.249Z"
+  }
+]
