--- packages/client/src/components/ViewPage/EachCardForCarousel1.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function EachCardForCarousel({ number }) {
-  return <div className="view-each-card-carousel">{number}</div>;
-}
